Corrects major error, minor typos and improves readability
There was a major error in the section introducing Higman's embedding
theorem. In particular, the term *finitely generated* was incorrectly
omitted when talking about the subgroups.
Moreover, a version of Higman's embedding lemma for infinitely generated
groups has been added as an appendix.

Some minor typos have been corrected.

Color has been added to the comments in the pseudocode in Section 4.

An appendix has been added which gives further exercises.